Time: 4 ms (42.18%), Space: 69.5 MB (79.64%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -3,19 +3,19 @@ class Solution {
         double ans=Integer.MIN_VALUE;
         int size=nums.length;
 
-        double sum=0;
+        int sum=0;
         for(int i=0;i<k;i++){
             sum=sum+nums[i];
         }
 
-        double avg=sum/k;
+        double avg=(double) sum/k;
         ans=Math.max(ans,avg);
         int l=0;
 
         for(int i=k;i<size;i++){
             sum=sum+nums[i]-nums[l];
             l++;
-            double new_avg=sum/k;
+            double new_avg=(double) sum/k;
          
             ans=Math.max(ans,new_avg);
         }
